service_backend/requirements.api.txt
gunicorn
uvicorn
fastapi
pydantic-settings
loguru
pymongo
matplotlib
pandas
numpy
mlflow
